Don't need --out anymore
I got tired of having to specify --out /dev/null to just see the
assembly for the program, so I made it not require it.

It outputs a warning if you leave it off, just as a reminder that you do
need it if you want to actually compile something.