Making some progress.
Cleaned up the code a bit, got rid of some warnings.
Added the init wait thing from the slides so that it loads the first
instruction, right now it waits two clock cycles and not just one,
because I was trying to get signaltap to cooperate, not that it did.